fix: update abstraction (add toYaml method, add YamlContext)
