+++ lib/src/yaml_context.dart
@@ -1,0 +1,10 @@
+class YamlContext {
+  final int indentSize;
+
+  final bool allowUnquotedStrings;
+
+  YamlContext({
+    required this.indentSize,
+    required this.allowUnquotedStrings,
+  });
+}
