+++ lib/yaml_writer.dart
@@ -1,4 +1,4 @@
 /// YAML Writer Library.
 library yaml_writer;
 
-export 'src/yaml_writer_base.dart';
+export 'src/yaml_writer.dart';
